Extracted inner classes from RuleBasedControllerBuilder and added them to dedicated package
